Airflow/requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
beautifulsoup4
pandas
requests
selenium
webdriver-manager
requests
python-dotenv
docling
mistralai